requirements.txt
Flask>=2.2
Flask-Login
Flask-WTF
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
WTForms
email-validator
Werkzeug
pandas
reportlab
